requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.114.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
mysql-connector-python==8.1.0
passlib==1.7.4
protobuf==4.21.12
pyasn1==0.6.1
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
python-dotenv==1.0.0
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.38.6
typing_extensions==4.14.1
uvicorn==0.23.2
watchfiles==1.1.0
websockets==15.0.1
